add packages with npm
add alert 03 feedback

+++ src/js/03-feedback.js
@@ -1,4 +1,4 @@
-import throttle from 'lodash.throttle';
+import * as throttle from 'lodash.throttle';
 
 const forma = document.querySelector('.feedback-form');
 const formData = JSON.parse(localStorage.getItem('feedback-form-state')) || {};
@@ -7,7 +7,7 @@ const { email = null, message = null } = formData;
 document.querySelector('input').value = email;
 document.querySelector('textarea').value = message;
 
-forma.addEventListener('input', _.throttle(saveLocalStorage, 500));
+forma.addEventListener('input', throttle(saveLocalStorage, 500));
 forma.addEventListener('submit', showConsole);
 
 function saveLocalStorage(event) {
@@ -18,6 +18,10 @@ function saveLocalStorage(event) {
 
 function showConsole(event) {
   event.preventDefault();
+  if (!email || !message) {
+    alert('Please, fulfill the forma');
+    return;
+  }
   console.log(formData);
   event.currentTarget.reset();
   localStorage.clear();
